feat: swipe-list on web
