+++ apps/next/package.json
@@ -52,6 +52,7 @@
     "remove-markdown": "^0.3.0",
     "resize-observer-polyfill": "^1.5.1",
     "setimmediate": "^1.0.5",
+    "swiper": "^8.4.2",
     "swr": "2.0.0-beta.6",
     "wagmi": "^0.6.0",
     "web3modal": "^1.9.8"
