+++ packages/app/lib/constants.ts
@@ -180,4 +180,5 @@ export const CURRENCY_NAMES = {
 
 export const HIDE_LINK_FOOTER_ROUTER_LIST = [
   "/nft/[chainName]/[contractAddress]/[tokenId]",
+  "/list",
 ];
